Personal Schedule changes
- logic moved to "profile" folder/controller
- "personal schedule" menu item is hidden for now
