+++ TeamSpark.AzureDay.WebSite.Host/TeamSpark.AzureDay.WebSite.Host.csproj
@@ -544,7 +544,7 @@
     <Content Include="Theme\fonts\glyphicons-halflings-regular.ttf" />
     <Content Include="Theme\fonts\glyphicons-halflings-regular.woff" />
     <Content Include="Theme\fonts\glyphicons-halflings-regular.woff2" />
-    <Content Include="Views\Home\PersonalSchedule.cshtml" />
+    <Content Include="Views\Profile\PersonalSchedule.cshtml" />
   </ItemGroup>
   <ItemGroup>
     <Folder Include="App_Data\" />
