+++ TeamSpark.AzureDay.WebSite.Host/Views/Shared/_MenuItems.cshtml
@@ -2,7 +2,7 @@
 
 <li><a href="/">@_MenuItems.Home</a></li>
 <li><a href="/schedule">@_MenuItems.Agenda</a></li>
-<li><a href="/personalschedule">@_MenuItems.PersonalSchedule</a></li>
+<li>@*<a href="/profile/personalschedule">@_MenuItems.PersonalSchedule</a>*@</li>
 <li><a href="/workshops">@_MenuItems.Workshops</a></li>
 @*<li><a href="/partners">@_MenuItems.Partners</a></li>*@
